server/requirement.txt
Django==1.11.13
django-cors-headers==2.3.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
mongoengine==0.15.0
pymongo==3.6.1
pytz==2018.4
six==1.11.0
